Store .py files for icon resources in the repository
This obviates PyQt5 and PySide as build-time dependencies, which avoids
an issue where this causes conda to make PyQt5 a run-time dependency,
and furthermore a pinned version of it (it must be assuming it needs ABI
compatibility as if we were building from Qt itself).
