+++ setup.cfg
@@ -28,5 +28,3 @@ pyside = PySide2
 
 [dist_conda]
 conda_name_differences = PyQt5:pyqt
-# For build requirements - PySide2 is only in conda-forge.
-channels = conda-forge
